+++ Moments/comments/location.blade.php
@@ -22,7 +22,7 @@
             @component('components.comment.list', [
                 'comment' => $comment,
                 'detailLink' => true,
-                'sectionCreatorLiked' => false,
+                'sectionAuthorLiked' => false,
             ])@endcomponent
         @endforeach
     </div>