+++ Moments/components/comment/sticky.blade.php
@@ -1,5 +1,5 @@
 @component('components.comment.list', [
     'comment' => $sticky,
     'detailLink' => true,
-    'sectionCreatorLiked' => true,
+    'sectionAuthorLiked' => true,
 ])@endcomponent