+++ Moments/hashtags/detail.blade.php
@@ -40,7 +40,7 @@
                     @component('components.comment.list', [
                         'comment' => $comment,
                         'detailLink' => true,
-                        'sectionCreatorLiked' => false,
+                        'sectionAuthorLiked' => false,
                     ])@endcomponent
                 @endforeach
             </div>