+++ Moments/posts/detail.blade.php
@@ -18,7 +18,7 @@
 
     {{-- Comment Box --}}
     @component('components.post.section.comment-box', [
-        'nickname' => $post['creator']['nickname'],
+        'nickname' => $post['author']['nickname'],
         'pid' => $post['pid'],
     ])@endcomponent
 
@@ -33,7 +33,7 @@
                     @component('components.comment.sticky', [
                         'sticky' => $sticky,
                         'detailLink' => true,
-                        'sectionCreatorLiked' => true,
+                        'sectionAuthorLiked' => true,
                     ])@endcomponent
                 @endforeach
             </div>
@@ -65,7 +65,7 @@
                 @component('components.comment.list', [
                     'comment' => $comment,
                     'detailLink' => true,
-                    'sectionCreatorLiked' => true,
+                    'sectionAuthorLiked' => true,
                 ])@endcomponent
             @endforeach
         </div>