+++ Moments/profile/dislikes/comments.blade.php
@@ -7,7 +7,7 @@
             @component('components.comment.list', [
                 'comment' => $comment,
                 'detailLink' => true,
-                'sectionCreatorLiked' => false,
+                'sectionAuthorLiked' => false,
             ])@endcomponent
         @endforeach
     </div>