+++ ThemeFrame/comments/following.blade.php
@@ -18,7 +18,7 @@
                         @component('components.comment.list', [
                             'comment' => $comment,
                             'detailLink' => true,
-                            'sectionCreatorLiked' => false,
+                            'sectionAuthorLiked' => false,
                         ])@endcomponent
 
                         @if (! $loop->last)