+++ ThemeFrame/hashtags/detail.blade.php
@@ -45,7 +45,7 @@
                                 @component('components.comment.list', [
                                     'comment' => $comment,
                                     'detailLink' => true,
-                                    'sectionCreatorLiked' => false,
+                                    'sectionAuthorLiked' => false,
                                 ])@endcomponent
 
                                 @if (! $loop->last)